LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/persisted-user-state-restoration/scroll-restoration-basic-expected.txt


PASS Default value is "auto"
PASS It is writable
PASS Invalid values are ignored
